requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.1
pyproject_hooks==1.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
WTForms==3.0.1
